requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
python-dotenv==1.0.0
supabase==1.0.3
werkzeug==2.3.7
boto3==1.28.62
botocore==1.31.62
requests==2.31.0